+++ src/lib/ic/libbuzztard-ic/abs-range-control.h
@@ -60,6 +60,6 @@ struct _BtIcAbsRangeControlClass {
 
 GType btic_abs_range_control_get_type(void) G_GNUC_CONST;
 
-BtIcAbsRangeControl *btic_abs_range_control_new(const BtIcDevice *device,const gchar *name,guint id,gint32 min,gint32 max,gint32 def);
+BtIcAbsRangeControl *btic_abs_range_control_new(const BtIcDevice *device,const gchar *name,guint id,glong min,glong max,glong def);
 
 #endif // BTIC_ABS_RANGE_CONTROL_H
